Add Qualif tests for iterator filters
This includes both decision and mcdc tests. Among the four constructs
in which iterator filters can be expected (array aggregates, container
aggregates, for-loops and quantified expressions), all are excercized
both for decision and MC/DC, but container aggregates and quantified
expressions tests are not implemented for all existing topologies, to
cut down on test driver modifications.