Fix the instrumentation of mains for C23
In C23 mode (`-std=c23`), Clang cannot parse statements that cannot be
completely typed.

Since the C instrumenters adds calls to witness functions coming from
the coverage runtime, this means that Clang needs to have access to the
coverage runtime headers in order to correctly parse the mains to
instrument.

For instance, the following unit:

```
int main (void) {
    ...
    return a && b;
}
```

Is instrumented into the following for MC/DC:

```
#include "gnatcov_rts_c-witness.h"
int main (void) {
    unsigned mcdc_state_var;
    unsigend *mcdc_state = &mcdc_state_var;
    gnatcov_rts_witness (...);
    return gnatcov_rts_witness_decision_mcdc(...);
}
```

The main instrumentation needs to have special processing for the return
statement, so that first the value to return is evaluated, and only then
the dump of coverage buffers happens. For this to work, Clang needs to
be able to parse the return statement. This was not the case so far,
because the `gnatcov_rts_c-witness.h` header was not necessarily
available during instrumentation.

Following the recent work to pass `-std=c23` to Clang when appropriate,
the `V824-009-dump-trigger-main-end/main-unit-of-interest` testcase
broke because of this.

To fix this case, this commit adds the coverage runtime source
directories to the setup configuration JSON file, and uses that
configuration whenever a C main is instrumented to make the coverage
runtime C headers available to Clang.