Improve the performance of consolidating source traces
When reading source traces, consolidate coverage buffers first (this is
quite fast), and only then process the consolidated buffers to update
gnatcov's coverage state in one pass (this is slow). This dramatically
improves the performance of `gnatcov coverage` when it is given a lot of
trace files. Using our performance benchmark on 'gnatcov' itself, the
performance of `gnatcov coverage` goes from:

* ~4.3 (1 trace)
* ~5s (500 traces)
* ~15s (10,000 traces)

to

* ~4.3 (1 trace)
* ~4.7s (500 traces)
* ~8.7s (10,000 traces)