coverage-source.adb: make SCI vectors store access rather than records
The `Source_Coverage_Info` record is complex: discriminated, with nested
containers. As a consequence, `SCI_Vectors.Vector` is costly to use: the
vector re-sizing incurs a big runtime overhead. Introduce an access
indirection so that the re-sizing implies a simple reallocation.

Using our performance benchmark on 'gnatcov' itself, the performance of
`gnatcov coverage` on 10,000 traces goes from ~7s to ~5.5s.