instrument-input_traces.adb: avoid string formatting when not needed
Calling `Unit_Image` in `Update_State` is slow, and generally not needed
(i.e. only when the `Misc_Trace` is active). Skip the call if not needed
to make `gnatcov coverage` faster.

Using our performance benchmark on 'gnatcov' itself, the performance of
`gnatcov coverage` on 10,000 traces goes from ~4.8s to ~4.6s.