Integrated instrumentation tests: fix support for cross target
When testing the integrated instrumentation approach for cross, using the
C BSP provided with the toolchain, compile the instrumentation runtime
specifically for the BSP, thus with the set of compiler and linker
switches that it provides. Otherwise, we end up using the default
compiler configuration, which may mismatch with the board that the BSP
targets.