Do not create MCDC state vars for non-instrumented decisions
MCDC state variables are useless if they are created for decisions that
are not instrumented eventually: do not generate them in the first
place for such decisions.

For now, this is just a quality of implementation improvement: these
varibles were just bloat. However, upcoming changes may create them in
declare expressions inside the decisions themselves, which can yield
illegal Ada code (for instance in the `Ada2022/subtype_ind_in_iter`
testcase).