Kill test on arm-elf
The test does not use gprfor to generate the gpr file, which is thus
missing the board information, ultimately resulting in a qemu failure in
the arm-elf configuration, which has two boards (and the wrong one ends
up being selected by default).

As this test only checks a deprecation warnings, kill it on arm-elf to
avoid testsuite adaptations.